+++ app/views/components/buttons/search_component.html.erb
@@ -1,0 +1,3 @@
+<button <%= tag.attributes(class: class_name, type: "button") %>>
+  <%= content %>
+</button>
