+++ app/views/components/navbars/top/link_item_component.html.erb
@@ -1,8 +1,8 @@
 <%= link_to path, {
-  class: class_names("text-gray-300 hover:text-gray-200", class_name)
+  class: class_names("text-gray-300 hover:text-gray-200 text-sm", class_name)
 } do %>
   <div class="flex items-center gap-2">
-    <%= render Basic::IconComponent.new(name: icon_name, size: "18px", class_name: "fill-gray-200") %>
+    <%= render Basic::IconComponent.new(name: icon_name, size: "16px", class_name: "fill-gray-200") %>
 
     <%= title %>
   </div>
