+++ app/views/components/navbars/top/search_button_component.rb
@@ -1,0 +1,9 @@
+# typed: strict
+# frozen_string_literal: true
+
+module Navbars
+  module Top
+    class SearchButtonComponent < ApplicationComponent
+    end
+  end
+end
