--- app/views/components/navbars/top/search_component.html.erb
@@ -1,6 +1,0 @@
-<input
-  autocomplete="off"
-  class="input rounded-full border-gray-400 bg-transparent text-gray-200 transition duration-200 ease-in-out focus:border-gray-200 h-auto py-2"
-  type="text"
-  placeholder="Search..."
->
