+++ app/views/components/navbars/top/user_component.html.erb
@@ -1,19 +1,20 @@
 <header class="<%= class_names("bg-base-200 drop-shadow min-h-full navbar py-2", class_name) %>">
-  <div class="navbar-start">
+  <div class="gap-12 navbar-start">
     <%= link_to root_path do %>
       <%= render Basic::IconComponent.new(name: "logo", size: "2rem", class_name: "fill-white") %>
     <% end %>
+
+    <div class="flex gap-6">
+      <%= render Navbars::Top::LinkItemComponent.new(path: root_path, title: t("nouns.home"), icon_name: "house") %>
+      <%= render Navbars::Top::LinkItemComponent.new(path: "#", title: t("nouns.inbox"), icon_name: "tray") %>
+    </div>
   </div>
 
-  <div class="gap-6 navbar-center">
-    <%= render Navbars::Top::LinkItemComponent.new(path: root_path, title: t("nouns.home"), icon_name: "house") %>
-    <%= render Navbars::Top::LinkItemComponent.new(path: "#", title: t("nouns.unreads"), icon_name: "rows") %>
-    <%= render Navbars::Top::LinkItemComponent.new(path: "#", title: t("nouns.notifications"), icon_name: "bell") %>
+  <div class="navbar-center">
+    <%= render Navbars::Top::SearchButtonComponent.new %>
   </div>
 
-  <div class="gap-6 navbar-end">
-    <%= render Navbars::Top::SearchComponent.new %>
-
+  <div class="navbar-end">
     <%= link_to "#" do %>
       <%= image_tag "https://placehold.co/300x300", class: "w-8 h-8 rounded-full" %>
     <% end %>
